msm: clock-debug: Replace ULONG_MAX with INT_MAX for 'fmax' assignment
The 'fmax' variable is an integer, not an unsigned long. Use the
correct macro so that the comparisons of 'rate' and 'fmax' works
correctly. Now, doing a 'cat list_rates' for a clock without voltage
constraints prints a non-empty list.

+++ arch/arm/mach-msm/clock-debug.c
@@ -174,7 +174,7 @@ static int list_rates_show(struct seq_file *m, void *unused)
 
 	/* Find max frequency supported within voltage constraints. */
 	if (!clock->vdd_class) {
-		fmax = ULONG_MAX;
+		fmax = INT_MAX;
 	} else {
 		for (level = 0; level < ARRAY_SIZE(clock->fmax); level++)
 			if (clock->fmax[level])
